+++ facebook_auth_web/lib/src/facebook_auth_plugin.dart
@@ -1,12 +1,14 @@
 import 'dart:async';
-import 'dart:html';
+import 'dart:js_interop';
+import 'dart:js_interop_unsafe';
+
 import 'package:flutter/services.dart';
 import 'package:flutter_facebook_auth_platform_interface/flutter_facebook_auth_platform_interface.dart';
-import 'package:js/js.dart';
-import 'dart:js' as js;
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
-import 'interop/facebook_auth_interop.dart' as fb;
+import 'package:web/web.dart';
+
 import 'interop/convert_interop.dart';
+import 'interop/facebook_auth_interop.dart' as fb;
 
 /// A web implementation of the FlutterFacebookAuth plugin.
 class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
@@ -27,15 +29,11 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
     if (!_initialized) return null;
 
     Completer<LoginResult> completer = Completer();
-    fb.getLoginStatus(
-      allowInterop(
-        (jsResponse) {
-          this._handleResponse(jsResponse).then(
-                (result) => completer.complete(result),
-              );
-        },
-      ),
-    );
+    fb.getLoginStatus((JSAny jsResponse) {
+      this._handleResponse(jsResponse).then(
+            (result) => completer.complete(result),
+          );
+    }.toJS);
     final LoginResult result = await completer.future;
     return result.accessToken;
   }
@@ -73,15 +71,14 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
     if (!_initialized) return {"error": "window.FB is undefined"};
     Completer<Map<String, dynamic>> c = Completer();
     fb.api(
-      "/me?fields=$fields",
-      allowInterop(
-        (_) => c.complete(
-          Map<String, dynamic>.from(
-            convert(_),
-          ),
-        ),
-      ),
-    );
+        "/me?fields=$fields",
+        (JSAny _) {
+          c.complete(
+            Map<String, dynamic>.from(
+              convert(_),
+            ),
+          );
+        }.toJS);
     return c.future;
   }
 
@@ -90,13 +87,11 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
   Future<void> logOut() async {
     if (!_initialized) return;
     Completer<void> c = Completer();
-    fb.logout(allowInterop(
-      (_) {
-        if (!c.isCompleted) {
-          c.complete();
-        }
-      },
-    ));
+    fb.logout((JSAny _) {
+      if (!c.isCompleted) {
+        c.complete();
+      }
+    }.toJS);
     return c.future;
   }
 
@@ -105,6 +100,8 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
   Future<LoginResult> login({
     List<String> permissions = const ['email', 'public_profile'],
     LoginBehavior loginBehavior = LoginBehavior.dialogOnly,
+    LoginTracking loginTracking = LoginTracking.enabled,
+    String? nonce,
   }) async {
     if (!_initialized) {
       return LoginResult(
@@ -115,13 +112,11 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
     String scope = permissions.join(",");
     Completer<LoginResult> completer = Completer();
     fb.login(
-      allowInterop(
-        (jsResponse) {
-          this._handleResponse(jsResponse).then(
-                (result) => completer.complete(result),
-              );
-        },
-      ),
+      (JSAny jsResponse) {
+        this._handleResponse(jsResponse).then(
+              (result) => completer.complete(result),
+            );
+      }.toJS,
       fb.LoginOptions(
         scope: scope,
         return_scopes: true,
@@ -135,7 +130,7 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
   ///
   /// calls the FB.init interop
   @override
-  Future<void> webInitialize({
+  Future<void> webAndDesktopInitialize({
     required String appId,
     required bool cookie,
     required bool xfbml,
@@ -143,7 +138,7 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
   }) async {
     this._appId = appId;
 
-    if (js.context['FB'] != null) {
+    if (globalContext.has("FB")) {
       _initialized = true;
       return;
     }
@@ -164,7 +159,7 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
   /// Injects a `script` with a `src` dynamically into the head of the current
   /// document.
   Future<void> _injectSrcScript() async {
-    final script = ScriptElement()
+    final script = HTMLScriptElement()
       ..type = 'text/javascript'
       ..src = 'https://connect.facebook.net/en_US/sdk.js'
       ..async = true
@@ -173,59 +168,6 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
     assert(document.head != null);
     document.head!.append(script);
     await script.onLoad.first;
-  }
-
-  /// get the granted and declined permission for the current facebook session
-  ///
-  /// The facebook SDK will return a JSON like
-  /// ```
-  /// {
-  ///  'data': [
-  ///    {
-  ///      "permission": "email",
-  ///      "status": "granted",
-  ///    },
-  ///    {
-  ///      "permission": "photos",
-  ///      "status": "declined",
-  ///    }
-  ///  ],
-  ///}
-  /// ```
-  @override
-  Future<FacebookPermissions?> get permissions async {
-    if (!_initialized) return null;
-    Completer<FacebookPermissions?> c = Completer();
-    fb.api(
-      "/me/permissions",
-      allowInterop(
-        (_) {
-          try {
-            List<String> granted = [];
-            List<String> declined = [];
-            final response = convert(_);
-            _checkResponseError(response);
-            for (final item in response['data'] as List) {
-              final String permission = item['permission'];
-              if (item['status'] == 'granted') {
-                granted.add(permission);
-              } else {
-                declined.add(permission);
-              }
-            }
-            c.complete(
-              FacebookPermissions(granted: granted, declined: declined),
-            );
-          } on PlatformException catch (e) {
-            print(
-              StackTrace.fromString(e.message ?? 'unknown error'),
-            );
-            c.complete(null);
-          }
-        },
-      ),
-    );
-    return c.future;
   }
 
   /// handle the login or getLoginStatus response
@@ -254,19 +196,23 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
       final String? status = response['status'];
 
       if (status == null) {
-        return LoginResult(status: LoginStatus.failed);
+        return LoginResult(
+          status: LoginStatus.failed,
+        );
       }
       if (status == 'connected') {
         final Map<String, dynamic> authResponse = response['authResponse'];
 
         final expires = DateTime.now().add(
-          Duration(seconds: authResponse['data_access_expiration_time']),
+          Duration(
+            seconds: authResponse['expiresIn'],
+          ),
         );
 
         // create a Login Result with an accessToken
         return LoginResult(
           status: LoginStatus.success,
-          accessToken: AccessToken(
+          accessToken: ClassicToken(
             applicationId: this._appId,
             grantedPermissions:
                 null, // on web we don't have this data in the login response
@@ -274,18 +220,21 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
                 null, // on web we don't have this data in the login response
             userId: authResponse['userID'],
             expires: expires,
-            lastRefresh: DateTime.now(),
-            token: authResponse['accessToken'],
-            isExpired: false,
-            graphDomain: authResponse['graphDomain'],
+            tokenString: authResponse['accessToken'],
           ),
         );
       } else if (status == 'unknown') {
         return LoginResult(status: LoginStatus.cancelled);
       }
-      return LoginResult(status: LoginStatus.failed, message: 'unknown error');
-    } on PlatformException catch (e) {
-      return LoginResult(status: LoginStatus.failed, message: e.message);
+      return LoginResult(
+        status: LoginStatus.failed,
+        message: 'unknown error',
+      );
+    } on PlatformException catch (e, _) {
+      return LoginResult(
+        status: LoginStatus.failed,
+        message: e.message,
+      );
     }
   }
 
@@ -293,7 +242,9 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
   void _checkResponseError(Map<String, dynamic> response) {
     if (response['error'] != null) {
       throw PlatformException(
-          code: "REQUEST_ERROR", message: response['error']['message']);
+        code: "REQUEST_ERROR",
+        message: response['error']['message'],
+      );
     }
   }
 
@@ -301,7 +252,9 @@ class FlutterFacebookAuthPlugin extends FacebookAuthPlatform {
   bool get isWebSdkInitialized => _initialized;
 
   @override
-  Future<void> autoLogAppEventsEnabled(bool enabled) async {}
+  Future<void> autoLogAppEventsEnabled(bool enabled) {
+    throw UnimplementedError();
+  }
 
   @override
   Future<bool> get isAutoLogAppEventsEnabled => Future.value(false);